requirements/requirements.txt
# CORE
Django>=3.1.0,<3.2.0
djangorestframework>=3.12.0,<3.13.0
djangorestframework-gis==0.17
django-leaflet==0.27.1
psycopg2-binary>=2.8.0,<2.9.0
python-dotenv>=0.15.0,<1.0.0
drf-spectacular>=0.17.0,<0.18.0
django-debug-toolbar>=3.2,<3.3


# SOCIAL AUTH
cryptography==3.1.1
drf-social-oauth2==1.0.9
django-oauth-toolkit>=1.4.0,<1.5.0
PyJWT == 1.7.1


# AWS related
# boto3>=1.17.99,<1.18.0
# django-storages>=1.11.1,<1.12.0


# DEV
# pytest-django>=4.2.0,<4.3
# flake8>=3.8.0,<3.9.0
# isort>=5.8.0,<5.9.0
# black
# factory_boy>=3.2.0,<3.3.0


# PROD
# gunicorn>=20.1.0,<20.2.0


# Previous Version
# geocoder~=1.38.1
# spotipy~=2.16.1
# pandas~=1.1.4
# scipy~=1.5.4
# requests~=2.24.0
# numpy~=1.19.3
# geopy~=2.0.0